django/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.1.0
bs4==0.0.1
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==0.7
cssselect==1.2.0
Django==4.2.5
django-crispy-forms==2.0
fake-useragent==1.2.1
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
kombu==5.3.2
lxml==4.9.3
multidict==6.0.4
numpy==1.25.2
packaging==23.1
pandas==2.1.0
parse==1.19.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pyee==8.2.2
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.0
requests==2.31.0
requests-html==0.10.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
w3lib==2.1.2
wcwidth==0.2.6
websockets==10.4
yarl==1.9.2
zipp==3.16.2
